// frontend/src/apiService/models/SessionStage.ts
/* tslint:disable */
/* eslint-disable */
/**
 * FastAPI
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Stages for coordinating host and audience clients
 * @export
 */
export const SessionStage = {
    JoinSession: 'join_session',
    AskQuestion: 'ask_question',
    AwaitAnswers: 'await_answers',
    ShowAnalyses: 'show_analyses'
} as const;
export type SessionStage = typeof SessionStage[keyof typeof SessionStage];


export function instanceOfSessionStage(value: any): boolean {
    for (const key in SessionStage) {
        if (Object.prototype.hasOwnProperty.call(SessionStage, key)) {
            if (SessionStage[key as keyof typeof SessionStage] === value) {
                return true;
            }
        }
    }
    return false;
}

export function SessionStageFromJSON(json: any): SessionStage {
    return SessionStageFromJSONTyped(json, false);
}

export function SessionStageFromJSONTyped(json: any, ignoreDiscriminator: boolean): SessionStage {
    return json as SessionStage;
}

export function SessionStageToJSON(value?: SessionStage | null): any {
    return value as any;
}

export function SessionStageToJSONTyped(value: any, ignoreDiscriminator: boolean): SessionStage {
    return value as SessionStage;
}